requirements/py38_testing.txt
appdirs==1.4.3
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
black==19.10b0
cffi==1.13.2
channels==2.4.0
Click==7.0
colorama==0.4.3
constantly==15.1.0
coverage==5.0.3
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
more-itertools==8.0.2
packaging==19.2
pathspec==0.7.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
pytest==5.3.2
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-django==3.7.0
pytz==2019.3
regex==2020.1.8
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.1
wcwidth==0.1.7
whitenoise==5.0.1
zope.interface==4.7.1
